+++ rally_openstack/common/cfg/neutron.py
@@ -27,7 +27,7 @@ OPTS = {"openstack": [
     cfg.BoolOpt("pre_newton_neutron",
                 default=False,
                 help="Whether Neutron API is older then OpenStack Newton or "
-                     "not. Based in this option, some external fields for "
+                     "not. Based on the option, some external fields for "
                      "identifying resources can be applied."),
     cfg.ListOpt("neutron_bind_l2_agent_types",
                 # default to agent types used in gate jobs